feat: Add CreatePost page component
The commit adds a new CreatePost page component to allow users to create new posts. It imports the necessary dependencies and renders a form for creating a new post. The form includes input fields for title and body. This component will be used to handle the creation of new posts in the application.